Solve day 17 using a BFS algorithm
Change the algorithm for day 17 to use a BFS rather than a DFS.

Like my original solution, this one uses caching for optimization. However, this solution does so, not by tagging a recursive function with `@cache`, but by caching the accumulated min heat loss for each position and direction as a vector, then pruning future paths that reach the same vector but exceed that vector's stored min heat loss value, whilst updating the value when a lower min is found.

The BFS algorithm is much faster than the DFS algorithm because it reaches a minimum heat loss closer to the optimum path much more quickly. This is because time is not wasted exploring deep branches of the tree that massively exceed the optimum path.

This solution solves both parts of the problem, in sequence, in 188 seconds.