Add bad solution to day 18 part 2
This solution is too bad/slow to handle the huge distances introduced by part 2. It takes forever to generate the matrix, suggesting that the use of a matrix to store every square meter of the build site (now thousands of km across) is not an optimal structure for this quantity of data.